Refactor game.js to import displayQuestion and displayAttempts from game-display.js
