+++ game.html
@@ -119,6 +119,7 @@
   <script type="module" src="assets/js/db.js"></script>
   <script type="module" src="assets/js/game-storage.js"></script>
   <script type="module" src="assets/js/random-object.js"></script>
+  <script type="module" src="assets/js/game-display.js"></script>
   <script type="module" src="assets/js/game.js"></script>
 </body>
